+++ dbt/adapters/firebolt/impl.py
@@ -196,27 +196,29 @@ class FireboltAdapter(SQLAdapter):
             )
 
     @available.parse_none
-    def get_datatype_table_from_SDKColumns(
+    def sdk_column_list_to_firebolt_column_list(
         self, columns: List[SDKColumn]
-    ) -> agate.Table:
+    ) -> List[FireboltColumn]:
         """
         Extract and return list of FireboltColumns with names and data types
         extracted from SDKColumns.
         Args:
           columns: list of Column types as defined in the Python SDK
         """
-        column_names = ['name', 'data_type']
-        column_types = [agate.Text(), agate.Text()]
-        for col in columns:
-            print(
-                f'\n\n** data types: {type(col)}, '
-                f'{col}, {col.type_code}, {col.type_code.__name__}'
-            )
-        rows = [{'name': col.name, 'data_type': col.type_code} for col in columns]
-
-        agate.Table(rows, column_names, column_types)
-        rows = [FireboltColumn(column=col.name, dtype=col.type_code) for col in columns]
-        return rows
+        types = {
+            'str': 'TEXT',
+            'int': 'INT',
+            'float': 'FLOAT',
+            'date': 'DATE',
+            'datetime': 'DATE',
+            'bool': 'BOOLEAN',
+            'list': 'ARRAY',
+            'Decimal': 'DECIMAL',
+        }
+        return [
+            FireboltColumn(column=col.name, dtype=types[col.type_code.__name__])
+            for col in columns
+        ]
 
     @available.parse_none
     def filter_table(
@@ -227,18 +229,6 @@ class FireboltAdapter(SQLAdapter):
         https://agate.readthedocs.io/en/latest/cookbook/filter.html#by-regex
         """
         return agate_table.where(lambda row: re.match(re_match_exp, str(row[col_name])))
-
-    @available.parse_none
-    def group_tuples(tuples: List[tuple]) -> List[tuple]:
-        """
-        Return a list of all unique tuples.
-        Args:
-         tuples: A list of tuples.
-        """
-        output = set()
-        for tp in tuples:
-            output.add(tp)
-        return list(output)
 
     @available.parse_none
     def get_rows_different_sql(
@@ -265,17 +255,43 @@ class FireboltAdapter(SQLAdapter):
             f'{relation_a}.{name} = {relation_b}.{name}' for name in names
         ]
         where_clause = ' AND '.join(where_expressions)
-
         columns_csv = ', '.join(names)
-
         sql = COLUMNS_EQUAL_SQL.format(
             columns=columns_csv,
             relation_a=str(relation_a),
             relation_b=str(relation_b),
             where_clause=where_clause,
         )
-
         return sql
+
+    @available.parse_none
+    def annotate_date_columns_for_partitions(
+        self,
+        vals: str,
+        col_names: Union[List[str], str],
+        col_types: List[FireboltColumn],
+    ) -> str:
+        """
+        Return a list of partition values as a single string. All columns with
+        date types will be be suffixed with ::DATE.
+        Args:
+          vals: a string of values separated by commas
+          col_names: either a list of strings or a single string, of the
+            names of the columns
+          col_types: Each FireboltColumn has fields for the name of the column
+            and its type.
+        """
+        vals_list = vals.split(',')
+        # It's possible that col_names will be single column, in which case
+        # it might come in as a string.
+        if type(col_names) is str:
+            col_names = [col_names]
+        # Now map from column name to column type.
+        type_dict = {c.name: c.dtype for c in col_types}
+        for i in range(len(vals_list)):
+            if type(type_dict[col_names[i]]) in ['datetime', 'date']:
+                vals_list[i] += '::DATE'
+        return ','.join(vals_list)
 
 
 COLUMNS_EQUAL_SQL = """