+++ CHANGELOG.md
@@ -8,8 +8,7 @@
 
 ### Under the hood
 
-- Added mypy to pre-commit tests.
-- Fixed output from dbt runs so that successful table creation outputs “Success” rather than “False.”
+- dbt-firebolt now requires firebolt-sdk>=0.8
 
 ## v.1.0.6
 
@@ -27,6 +26,8 @@
 ### Under the hood
 
 - Added FireboltColumn class to handle string type.
+- Added mypy to pre-commit tests.
+- Fixed output from dbt runs so that successful table creation outputs “Success” rather than “False.”
 
 ## v.1.0.4